requirements.txt
slack_bolt>=1.18.0
python-dotenv>=1.0.0
watchfiles>=0.21.0
anthropic>=0.25.0
requests>=2.31.0
claude-code-sdk>=0.0.25
pyyaml>=6.0.0
rapidfuzz>=3.0.0
whoosh>=2.7.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
